Move `File` and `DirectoryIterator` to separate files
