+++ Sts1CobcSw/FileSystem/CMakeLists.txt
@@ -1,4 +1,4 @@
-target_sources(Sts1CobcSw_FileSystem PRIVATE FileSystem.cpp)
+target_sources(Sts1CobcSw_FileSystem PRIVATE DirectoryIterator.cpp FileSystem.cpp File.cpp)
 target_link_libraries(
     Sts1CobcSw_FileSystem PUBLIC etl::etl littlefs::littlefs Sts1CobcSw_Outcome Sts1CobcSw_Serial
                                  Sts1CobcSw_Vocabulary
