+++ Sts1CobcSw/FileSystem/File.ipp
@@ -1,6 +1,7 @@
 #pragma once
 
-#include <Sts1CobcSw/FileSystem/FileSystem.hpp>
+
+#include <Sts1CobcSw/FileSystem/File.hpp>
 
 
 namespace sts1cobcsw::fs
